requirements.txt
cryptography==42.0.8
ddddocr==1.5.5
Flask==3.0.3
loguru==0.7.2
Pillow==10.4.0
pycryptodome==3.20.0
python_crontab==3.2.0
Requests==2.32.3
ruamel.base==1.0.0
selenium==3.8.0
SQLAlchemy==2.0.31
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
